adding act title scene and first person scene
