--- Storyweaver/FirstPersonScene.swift
@@ -1,27 +1,0 @@
-//
-//  FirstPersonScene.swift
-//  Storyweaver
-//
-//  Created by Gregorius Yuristama Nugraha on 6/20/23.
-//
-
-import SpriteKit
-import GameplayKit
-
-
-class FirstPersonScene: SKScene {
-    
-    var dialogueLabel: SKLabelNode!
-    
-    
-    
-    override init() {
-        super.init()
-        
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-}
