+++ Storyweaver/GameViewController.swift
@@ -17,16 +17,33 @@ class GameViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        // Create the SKView
-        gameView = SKView(frame: view.bounds)
-        view.addSubview(gameView)
+//        // Create the SKView
+//        gameView = SKView(frame: view.bounds)
+//        view.addSubview(gameView)
+//
+//        // Create the GameScene
+//        gameScene = GameScene(size: gameView.bounds.size)
+//        gameScene.scaleMode = .aspectFill
+//
+//        // Present the GameScene in the SKView
+//        gameView.presentScene(gameScene)
         
-        // Create the GameScene
-        gameScene = GameScene(size: gameView.bounds.size)
-        gameScene.scaleMode = .aspectFill
-        
-        // Present the GameScene in the SKView
-        gameView.presentScene(gameScene)
+        if let view = self.view as! SKView? {
+          // Load the SKScene from 'GameScene.sks'
+          if let scene = ActTitleScene(fileNamed: "ActTitleScene") {
+            // Set the scale mode to scale to fit the window
+            scene.scaleMode = .aspectFill
+                  
+            // Present the scene
+            view.presentScene(scene)
+          }
+
+//          view.ignoresSiblingOrder = true
+
+//          view.showsFPS = true
+//          view.showsNodeCount = true
+          view.ignoresSiblingOrder = false
+        }
     }
     
     override func viewDidAppear(_ animated: Bool) {
